requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
pymongo==4.6.1
pymongo[srv]==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
whitenoise==6.6.0
gunicorn==21.2.0
